move interpolation to new file
